Working stream - next get config from post.
