+++ .travis.yml
@@ -2,12 +2,10 @@ sudo: required
 
 language: python
 python:
-  -  "3.5"
+  -  "3.8"
 
 services:
   - docker
 
 script: true
 
-after_success:
-  - bash <(curl -s https://raw.githubusercontent.com/sesam-community/guidelines/master/after_success.sh)
